feat(installer): report the Windows Agent user check on the rc state
Move the rc state collection out of daemon.go into state.go, with state_windows.go
and state_nix.go for the platform-specific part.

On Windows, an Agent running as a domain account with no stored password can never
take a remote update -- it fails with installer error code 6. refreshState now runs
the Agent user check and reports its outcome as the installer error code an update
would fail with, so a host that cannot succeed can be left out of a rollout rather
than counted as a failed job.

  installer_agent_user_error_code:ok   check passed
  installer_agent_user_error_code:6    no stored password for a domain account
  installer_agent_user_error_code:0    check could not conclude

Only code 6 is deterministic. The others cover the SID lookup and gMSA query,
which need a reachable domain controller, so only 6 should be gated on. Absence
of the tag means unknown: not Windows, or an Agent that predates this.
